add pay button and order id

+++ app/src/main/java/com/example/keviniswara/bookinglapang/model/Order.kt
@@ -7,7 +7,7 @@ class Order(var customerEmail: String,
             var sport: String,
             var startHour: String,
             var status: Int,
-            var deadline: Int,
+            var deadline: Long,
             var orderId: String) {
     constructor() : this("", "", "", "", "", "",
             -1, 0, "")
